myapi/requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
djangorestframework==3.15.2
numpy==2.2.0
packaging==24.2
pandas==2.2.3
pillow==11.0.0
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
six==1.17.0
sqlparse==0.5.3
tzdata==2024.2
